Deleted activities, StringKeys related to intents
