--- app/src/main/java/com/tetsoft/typego/ui/activity/AccountActivity.kt
@@ -1,34 +1,0 @@
-package com.tetsoft.typego.ui.activity
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import android.view.View
-import android.widget.Toast
-import androidx.navigation.ActivityNavigator
-import androidx.navigation.findNavController
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import com.tetsoft.typego.R
-import com.tetsoft.typego.databinding.ActivityAccountBinding
-import com.tetsoft.typego.utils.StringKeys
-
-class AccountActivity : AppCompatActivity() {
-    private lateinit var binding : ActivityAccountBinding
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityAccountBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        val navHost =
-            supportFragmentManager.findFragmentById(R.id.fragmentContainerView) as NavHostFragment
-        binding.bottomNavigationView.setupWithNavController(navHost.navController)
-
-        binding.tvAccountActivityTitle.setOnClickListener { finish() }
-    }
-
-    override fun finish() {
-        super.finish()
-        ActivityNavigator.applyPopAnimationsToPendingTransition(this)
-    }
-}
