--- app/src/main/res/layout/activity_account.xml
@@ -1,64 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.activity.AccountActivity">
-
-    <LinearLayout
-        android:id="@+id/my_profile_action_bar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintTop_toTopOf="parent"
-        android:orientation="vertical"
-        android:background="@color/bg_dark_gray"
-        android:layout_marginBottom="4dp">
-
-        <TextView
-            android:id="@+id/tv_account_activity_title"
-            style="@style/activity_title"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp"
-            android:text="@string/button_Profile"
-            android:textColor="@color/text_accent"
-            app:drawableStartCompat="@drawable/ic_arrow_back"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
-
-        <View
-            android:id="@+id/account_activity_title_divider"
-            android:layout_width="match_parent"
-            android:layout_height="0.2dp"
-            android:background="@color/white"
-            app:layout_constraintTop_toBottomOf="@id/tv_account_activity_title" />
-
-    </LinearLayout>
-
-
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id="@+id/bottomNavigationView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/bg_dark_gray"
-        app:itemBackground="@drawable/bottom_nav_item"
-        app:itemIconTint="@color/white"
-        app:itemTextColor="@color/white"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:menu="@menu/bottom_menu" />
-
-    <androidx.fragment.app.FragmentContainerView
-        android:id="@+id/fragmentContainerView"
-        android:name="androidx.navigation.fragment.NavHostFragment"
-        android:layout_width="0dp"
-        android:layout_height="0dp"
-        app:defaultNavHost="true"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/my_profile_action_bar"
-        app:layout_constraintBottom_toTopOf="@id/bottomNavigationView"/>
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
