--- app/src/main/res/layout/activity_release_notes.xml
@@ -1,193 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView
-    android:id="@+id/releases_scroll"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    style="@style/scrollbars">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_margin="16dp"
-        android:orientation="vertical">
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            style="@style/basic_cardview">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="16dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:id="@+id/tv_whats_new"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/submenu_header"
-                    android:text="@string/whats_new_pl"
-                    app:drawableTint="@color/main_green"
-                    app:drawableStartCompat="@drawable/ic_release_notes"
-                    android:drawablePadding="4dp"
-                    android:layout_marginBottom="16dp"/>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fontFamily="@font/redring_regular"
-                    android:textColor="@color/white"
-                    android:text="@string/release_notes_description_current_version"/>
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/previous_versions"
-
-            android:layout_marginTop="24dp"
-            android:layout_marginBottom="8dp"
-            android:layout_gravity="center_horizontal"/>
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            style="@style/basic_cardview">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="16dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/submenu_header"
-                    android:text="1.4.8"
-                    app:drawableTint="@color/main_green"
-                    app:drawableStartCompat="@drawable/ic_release_notes"
-                    android:drawablePadding="4dp"
-                    android:layout_marginBottom="16dp"/>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fontFamily="@font/redring_regular"
-                    android:textColor="@color/white"
-                    android:text="@string/release_notes_description_148"/>
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            style="@style/basic_cardview">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="16dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/submenu_header"
-                    android:text="1.4.7"
-                    app:drawableTint="@color/main_green"
-                    app:drawableStartCompat="@drawable/ic_release_notes"
-                    android:drawablePadding="4dp"
-                    android:layout_marginBottom="16dp"/>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fontFamily="@font/redring_regular"
-                    android:textColor="@color/white"
-                    android:text="@string/release_notes_description_147"/>
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            style="@style/basic_cardview">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="16dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/submenu_header"
-                    android:text="1.4.3"
-                    app:drawableTint="@color/main_green"
-                    app:drawableStartCompat="@drawable/ic_release_notes"
-                    android:drawablePadding="4dp"
-                    android:layout_marginBottom="16dp"/>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fontFamily="@font/redring_regular"
-                    android:textColor="@color/white"
-                    android:text="@string/release_notes_description_143"/>
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-        <androidx.cardview.widget.CardView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginBottom="16dp"
-            style="@style/basic_cardview">
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:padding="16dp"
-                android:orientation="vertical">
-
-                <TextView
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    style="@style/submenu_header"
-                    android:text="1.4.0"
-                    app:drawableTint="@color/main_green"
-                    app:drawableStartCompat="@drawable/ic_release_notes"
-                    android:drawablePadding="4dp"
-                    android:layout_marginBottom="16dp"/>
-
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:fontFamily="@font/redring_regular"
-                    android:textColor="@color/white"
-                    android:text="@string/release_notes_description_140"/>
-
-            </LinearLayout>
-
-        </androidx.cardview.widget.CardView>
-
-    </LinearLayout>
-
-</androidx.core.widget.NestedScrollView>
-
