--- app/src/main/res/layout/activity_typing_test.xml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.activity.TypingTestActivity">
-
-    <com.tetsoft.typego.ui.custom.SpannableEditText
-        android:id="@+id/words"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_marginBottom="12dp"
-        android:background="@color/bg_dark_gray"
-        android:clickable="false"
-        android:focusable="false"
-        android:gravity="top|start"
-        android:inputType="textMultiLine"
-        android:padding="8dp"
-        android:textSize="30sp"
-        app:layout_constraintBottom_toTopOf="@id/inpWord"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <EditText
-        android:id="@+id/inpWord"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="8dp"
-        android:layout_marginBottom="8dp"
-        android:background="@drawable/text_input"
-        android:hint="@string/input_text_above"
-        android:imeOptions="flagNoExtractUi"
-        android:importantForAutofill="no"
-        android:inputType="textPersonName"
-        android:paddingStart="10dp"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@id/tvTimeLeft"
-        app:layout_constraintStart_toStartOf="parent" />
-
-    <TextView
-        android:id="@+id/tvTimeLeft"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="top"
-        android:text="@string/time_over"
-        android:textSize="24sp"
-        app:layout_constraintBottom_toBottomOf="@id/inpWord"
-        app:layout_constraintEnd_toStartOf="@id/tvTimeLeft"
-        app:layout_constraintTop_toTopOf="@+id/inpWord" />
-
-    <ImageButton
-        android:id="@+id/restartButton"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_gravity="top"
-        android:layout_marginHorizontal="8dp"
-        android:background="@null"
-        android:contentDescription="@string/content_desc_test_restart"
-        android:focusable="false"
-        android:padding="8dp"
-        android:scaleType="centerCrop"
-        android:src="@android:drawable/ic_popup_sync"
-        app:layout_constraintBottom_toBottomOf="@+id/tvTimeLeft"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@id/tvTimeLeft"
-        app:layout_constraintTop_toTopOf="@+id/tvTimeLeft" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
