+++ app/src/main/res/layout/fragment_game_setup.xml
@@ -5,7 +5,7 @@
     android:id="@+id/ConstraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activity.TestSetupActivity">
+    tools:context=".ui.fragment.setup.GameSetupFragment">
 
     <LinearLayout
         android:id="@+id/custom_test_action_bar"
