+++ app/src/main/res/layout/fragment_result.xml
@@ -5,7 +5,7 @@
     android:id="@+id/result_constraint_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.activity.ResultActivity">
+    tools:context=".ui.fragment.result.ResultFragment">
 
     <androidx.core.widget.NestedScrollView
         style="@style/scrollbars"
